+++ gstreamer/factories/TEXT/event/publishTrueInfo.cc
@@ -11,7 +11,7 @@ bool GstreamerTextFactory::publishEventTrueInfoData(const string detectorName, c
 
 		string identifierString = "";
 		vector<GIdentifier> gidentity = trueInfoHit->getIdentity();
-		for ( int i=0; i<gidentity.size() - 1; i++ ) {
+		for ( size_t i=0; i<gidentity.size() - 1; i++ ) {
 			identifierString += gidentity[i].getName() + "->" + to_string(gidentity[i].getValue()) + ", ";
 		}
 		identifierString += gidentity.back().getName() + "->" + to_string(gidentity.back().getValue()) ;
